requirements.txt
streamlit
PyPDF2
pandas
mistralai
python-dotenv
pillow
plotly
matplotlib
streamlit-agraph
pyautogen==0.4.1
duckduckgo_search==7.0.2
apify_client==1.8.1
flaml[automl]
snowflake==1.0.2
snowflake-snowpark-python==1.26.0
llmlingua==0.2.2
embedchain
youtube-transcript-api>=0.6.1
chromadb>=0.4.18
requests>=2.31.0
trulens-eval
trulens-connectors-snowflake==1.3.2
arxiv==2.1.3
pysqlite3-binary>=0.5.1 
trulens-connectors-snowflake
trulens-providers-openai